Extend plugin_sources list rather then overriding with the latest found (#268)
