+++ cibyl/plugins/__init__.py
@@ -52,9 +52,8 @@ def extend_source(plugin_name, plugin_module_path):
                if f.endswith('.py') and f != '__init__.py']:
         mod = __import__('.'.join(
             [f"cibyl.plugins.{plugin_name}.sources", py]), fromlist=[py])
-        new_element = [source_tuple[1] for source_tuple in
-                       inspect.getmembers(mod, is_plugin_class)]
-        plugin_sources.extend(new_element)
+        plugin_sources.extend([source_tuple[1] for source_tuple in
+                               inspect.getmembers(mod, is_plugin_class)])
     for plugin_source in plugin_sources:
         SourceFactory.extend_source(plugin_source)
