+++ cibyl/sources/elasticsearch/api.py
@@ -27,17 +27,15 @@ from cibyl.models.attribute import AttributeDictValue
 from cibyl.models.ci.base.build import Build
 from cibyl.models.ci.base.job import Job
 from cibyl.models.ci.base.test import Test
-from cibyl.plugins.openstack.deployment import Deployment
 from cibyl.sources.elasticsearch.client import ElasticSearchClient
 from cibyl.sources.server import ServerSource
 from cibyl.sources.source import speed_index
-from cibyl.utils.filtering import IP_PATTERN
 
 LOG = logging.getLogger(__name__)
 
 
-class ElasticSearchOSP(ServerSource):
-    """Used to perform queries in elasticsearch"""
+class ElasticSearch(ServerSource):
+    """Elasticsearch Source"""
 
     def __init__(self, driver: str = 'elasticsearch',
                  name: str = "elasticsearch", priority: int = 0,
@@ -208,174 +206,6 @@ class ElasticSearchOSP(ServerSource):
             job_object[job_name].add_build(build_object)
 
         return AttributeDictValue("jobs", attr_type=Job, value=job_object)
-
-    @speed_index({'base': 2})
-    def get_deployment(self, **kwargs):
-        """Get deployment information for jobs from elasticsearch server.
-
-        :returns: container of jobs with deployment information from
-        elasticsearch server
-        :rtype: :class:`AttributeDictValue`
-        """
-        jobs_found = self.get_jobs(**kwargs)
-
-        query_body = {
-            "query": {
-              "bool": {
-                "must": [
-                  {
-                    "bool": {
-                      "must": []
-                    }
-                  },
-                  {
-                    "bool": {
-                      "should": [
-                        {
-                          "exists": {
-                            "field": "ip_version"
-                          }
-                        },
-                        {
-                          "exists": {
-                            "field": "storage_backend"
-                          }
-                        },
-                        {
-                          "exists": {
-                            "field": "network_backend"
-                          }
-                        },
-                        {
-                          "exists": {
-                            "field": "dvr"
-                          }
-                        },
-                        {
-                          "exists": {
-                            "field": "topology"
-                          }
-                        }
-                      ],
-                      "minimum_should_match": 1
-                    }
-                  }
-                ]
-              }
-            },
-            "size": 1,
-            "sort": [
-                {
-                    "timestamp.keyword": {
-                        "order": "desc"
-                    }
-                }
-            ]
-        }
-
-        results = []
-        hits = []
-        for job in jobs_found:
-            query_body['query']['bool']['must'][0]['bool']['must'] = {
-                "match": {
-                    "job_name.keyword": f"{job}"
-                }
-            }
-            results = self.__query_get_hits(
-                query=query_body,
-                index='logstash_jenkins'
-            )
-            if results:
-                hits.append(results[0])
-
-        if not results:
-            return jobs_found
-
-        ip_version_argument = None
-        if 'ip_version' in kwargs:
-            ip_version_argument = kwargs.get('ip_version').value
-        dvr_argument = None
-        if 'dvr' in kwargs:
-            dvr_argument = kwargs.get('dvr').value
-        release_argument = None
-        if 'release' in kwargs:
-            release_argument = kwargs.get('release').value
-        network_argument = None
-        if 'network_backend' in kwargs:
-            network_argument = kwargs.get('network_backend').value
-        storage_argument = None
-        if 'storage_backend' in kwargs:
-            storage_argument = kwargs.get('storage_backend').value
-        if 'osp_release' in kwargs:
-            storage_argument = kwargs.get('osp_release').value
-
-        job_objects = {}
-        for hit in hits:
-            job_name = hit['_source']['job_name']
-            job_url = re.compile(r"(.*)/\d").search(
-                hit['_source']['build_url']
-            ).group(1)
-
-            # If the key exists assign the value otherwise assign unknown
-            topology = hit['_source'].get(
-                "topology", "unknown")
-            network_backend = hit['_source'].get(
-                "network_backend", "unknown")
-            ip_version = hit['_source'].get(
-                "ip_version", "unknown")
-            storage_backend = hit['_source'].get(
-                "storage_backend", "unknown")
-            dvr = hit['_source'].get(
-                "dvr", "unknown")
-            osp_release = hit['_source'].get(
-                "osp_release", "unknown")
-
-            if ip_version != 'unknown':
-                matches = IP_PATTERN.search(ip_version)
-                ip_version = matches.group(1)
-
-            # Check if necessary filter by IP version:
-            if ip_version_argument and \
-                    ip_version not in ip_version_argument:
-                continue
-
-            # Check if necessary filter by dvr:
-            if isinstance(dvr_argument, list) and \
-                    dvr not in dvr_argument:
-                continue
-
-            # Check if necessary filter by release version:
-            if release_argument and \
-                    osp_release not in release_argument:
-                continue
-
-            # Check if necessary filter by network backend:
-            if network_argument and \
-                    network_backend not in network_argument:
-                continue
-
-            # Check if necessary filter by storage backend:
-            if storage_argument and \
-                    storage_backend not in storage_argument:
-                continue
-
-            job_objects[job_name] = Job(name=job_name, url=job_url)
-            deployment = Deployment(
-                release=osp_release,
-                infra_type='',
-                nodes={},
-                services={},
-                ip_version=ip_version,
-                topology=topology,
-                network_backend=network_backend,
-                dvr=dvr,
-                storage_backend=storage_backend,
-                tls_everywhere=''
-            )
-
-            job_objects[job_name].add_deployment(deployment)
-
-        return AttributeDictValue("jobs", attr_type=Job, value=job_objects)
 
     @speed_index({'base': 3})
     def get_tests(self, **kwargs):
